Logica de permiso agregada a sistemaFCA

+++ app/Http/Controllers/AcademicoEventoController.php
@@ -8,43 +8,24 @@ use App\Evento;
 use Illuminate\Http\Request;
 use App\Http\Requests\AcademicoEventoRequest;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Session;
 
 class AcademicoEventoController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
-        return view('academicoEvento.index', [
-            'academicoEvento' => AcademicoEvento::get()
-        ]);
+        //
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
-        return view('academicoEvento.create', [
-            'eventoes' => Evento::get(),
-            'academicoes' => Academico::get()
-        ]);
+        //
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
+        Gate::authorize('havepermiso', 'participantes-crear');
         $request->validate([
             'academico' => 'required | numeric',
         ]);
@@ -59,7 +40,7 @@ class AcademicoEventoController extends Controller
 
         try {
             DB::beginTransaction();
-                DB::table('academico_evento')->insert([
+                DB::table('Academico_Evento')->insert([
                     'IdAcademico'       => $request['academico'],
                     'IdEvento'          => $request['evento']
                 ]);
@@ -73,47 +54,23 @@ class AcademicoEventoController extends Controller
         return redirect()->route('eventos.show', $request['evento']);
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Academico  $academico
-     * @return \Illuminate\Http\Response
-     */
     public function show(Academico $academico)
     {
         //
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Academico  $academico
-     * @return \Illuminate\Http\Response
-     */
     public function edit(Academico $academico)
     {
         //
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Academico  $academico
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request, Academico $academico)
     {
         //
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Academico  $academico
-     * @return \Illuminate\Http\Response
-     */
     public function destroy($participante) {
+        Gate::authorize('havepermiso', 'participantes-eliminar');
         $participante = AcademicoEvento::findOrFail( $participante );
         $participante->forceDelete();
